+++ resources/views/admin/categories/materials/partials/form.blade.php
@@ -92,5 +92,3 @@
 
     </div>
 </div>
-
-<style>[x-cloak]{display:none!important;}</style>