+++ resources/views/study/play.blade.php
@@ -270,9 +270,6 @@
     </div>
 </div>
 
-<style>
-    [x-cloak] { display: none !important; }
-</style>
 @endsection
 
 @section('js')